requirements.txt
requests
python-dotenv
chromadb
crewai
crewai-tools
langchain-google-genai
google-generativeai
pytest
